ci: disable duplicate build skipping (#360)
Build skipping has caused some problems when combined with dependabot
and merge protections. Activity in the C3R repo is at a level where
allowing occasional duplicate builds is not anticipated to cause issues
and is the simplest solution to the interfering actions.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.