feat: Add methods to get encoded bytes off of all `Value` based classes (#433)
*Description of changes:*
Add `getEncodedBytes()` to `Value` in addition to the existing
`getBytes()`. Encoded bytes are used for when a value will be encrypted
as it contains all metadata needed to reconstruct and verify the value.
Each format type will be responsible for encoding its own values because
different format implementations of `Value` may back one or many
different client data types.


By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.